Split up TestEGLNativeTypeWayland.cpp into separate translation units.
Where we don't generate the xbmc_wayland test extension, those tests
depending on it are disabled
